+++ src/main/java/walkingkooka/spreadsheet/dominokit/ui/find/SpreadsheetFindDialogComponentContext.java
@@ -21,8 +21,8 @@ import walkingkooka.datetime.HasNow;
 import walkingkooka.spreadsheet.dominokit.history.HistoryTokenContext;
 import walkingkooka.spreadsheet.dominokit.net.HasSpreadsheetDeltaFetcher;
 import walkingkooka.spreadsheet.dominokit.ui.dialog.SpreadsheetDialogComponentContext;
-import walkingkooka.spreadsheet.format.SpreadsheetParserProvider;
 import walkingkooka.spreadsheet.meta.HasSpreadsheetMetadata;
+import walkingkooka.spreadsheet.parser.SpreadsheetParserProvider;
 
 /**
  * The {@link walkingkooka.Context} accompanying a {@link SpreadsheetFindDialogComponent}.
